Support for two alternate screen ratio types.
Add the screenRatio property in a stage to change the size of the playable area.
There are still details to be tweaked about this playable area size thing, but right now it works for stages, at least.
The valid values right now are "narrow", "wide", and nil. Unrecognized values are treated as nil, and nil is the current, default playable area.